Create index.html
First version of the game. This is designed with Sybil attacks in mind. The first version is a "couch game" where everything is seen on one screen. This is totally done in vanillaJS with the understanding that some elements will be moved over into Solidit-- thus the use of the Keccak256 hash from ethers.